docs: remove old docs and install Vitepress
Old docs used Quartz, which is very impractical to use as a developer, because it is not an NPM package but a suite of files that one needs to update manually (at least when I started using it).

I'll now switch to something that I'm more familiar with for having used it on other projects as well.