docs: switch to Vitepress (#109)
Having a documentation based on an Obsidian vault is cool and practical when it comes to managing links and references between pages, however I find the engine I used to render the website a bit unpractical. The main reason is that it doesn't come as an NPM package, so it needs to be updated manually from time to time, meaning updating all code files manually.

It is doable, there are tools, but I just want something that works for this project, since I do not have much time to spend on it. Thus, I'm switching to Vitepress. I also think it ends up better-looking and easier to read.