refactor: move create notebooks feature to new code structure
Refactoring the whole plugin to make its codebase more modular and simpler to work with in terms of adding features, testing and fixing.

Deleting from the "old" code as I move the code to the new organization to avoid duplicated lines as much as possible while keeping the plugin working as before.