refactor: partially move checkpoints deletion feature to new code structure
Partially because I need to use an async event emitter, and I want to have a clean slate before trying it out.

Refactoring the whole plugin to make its codebase more modular and simpler to work with in terms of adding features, testing and fixing.

Deleting from the "old" code as I move the code to the new organization to avoid duplicated lines as much as possible while keeping the plugin working as before.