refactor: organize codebase (#117)
* refactor: move core features to new code structure

Refactoring the whole plugin to make its codebase more modular and simpler to work with in terms of adding features, testing and fixing.

* refactor: move create notebooks feature to new code structure

Refactoring the whole plugin to make its codebase more modular and simpler to work with in terms of adding features, testing and fixing.

Deleting from the "old" code as I move the code to the new organization to avoid duplicated lines as much as possible while keeping the plugin working as before.

* refactor: move update modal feature to new code structure

Refactoring the whole plugin to make its codebase more modular and simpler to work with in terms of adding features, testing and fixing.

Deleting from the "old" code as I move the code to the new organization to avoid duplicated lines as much as possible while keeping the plugin working as before.

* refactor: partially move checkpoints deletion feature to new code structure

Partially because I need to use an async event emitter, and I want to have a clean slate before trying it out.

Refactoring the whole plugin to make its codebase more modular and simpler to work with in terms of adding features, testing and fixing.

Deleting from the "old" code as I move the code to the new organization to avoid duplicated lines as much as possible while keeping the plugin working as before.

* feat: async event handling in Jupyter environment

* refactor: done moving code to new structure

* test: fix test taking current settings instead of default

* docs: document all refactored code and small cleanups